% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache.sedona-package.R
\docType{package}
\name{apache.sedona-package}
\alias{apache.sedona}
\alias{apache.sedona-package}
\title{apache.sedona: R Interface for Apache Sedona}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R interface for 'Apache Sedona' based on 'sparklyr' (\url{https://sedona.apache.org}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apache/sedona/}
  \item \url{https://sedona.apache.org/}
  \item Report bugs at \url{https://github.com/apache/sedona/issues}
}

}
\author{
\strong{Maintainer}:  Apache Sedona \email{private@sedona.apache.org}

Authors:
\itemize{
  \item Yitao Li \email{yitao@rstudio.com} (\href{https://orcid.org/0000-0002-1261-905X}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Jia Yu \email{jiayu@apache.org} [contributor, copyright holder]
  \item  The Apache Software Foundation [copyright holder]
  \item  RStudio [copyright holder]
}

}
\keyword{internal}
