% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rdd.R
\name{sedona_spatial_rdd_aggregation_routine}
\alias{sedona_spatial_rdd_aggregation_routine}
\title{Spatial RDD aggregation routine}
\arguments{
\item{x}{A Sedona spatial RDD.}
}
\description{
Function extracting aggregate statistics from a Sedona spatial RDD.
}
\keyword{internal}
