% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_write_wkb}
\alias{sedona_write_wkb}
\alias{sedona_write_wkt}
\alias{sedona_write_geojson}
\title{Write SpatialRDD into a file.}
\usage{
sedona_write_wkb(x, output_location)

sedona_write_wkt(x, output_location)

sedona_write_geojson(x, output_location)
}
\arguments{
\item{x}{The SpatialRDD object.}

\item{output_location}{Location of the output file.}
}
\value{
No return value.
}
\description{
Export serialized data from a Sedona SpatialRDD into a file.
\itemize{
\item \code{sedona_write_wkb}:
\item \code{sedona_write_wkt}:
\item \code{sedona_write_geojson}:
}
}
\examples{
library(sparklyr)
library(apache.sedona)

sc <- spark_connect(master = "spark://HOST:PORT")

if (!inherits(sc, "test_connection")) {
  input_location <- "/dev/null" # replace it with the path to your input file
  rdd <- sedona_read_wkb(
    sc,
    location = input_location,
    wkb_col_idx = 0L
  )
  sedona_write_wkb(rdd, "/tmp/wkb_output.tsv")
}

}
\seealso{
Other Sedona RDD data interface functions: 
\code{\link{sedona_read_dsv_to_typed_rdd}()},
\code{\link{sedona_read_geojson}()},
\code{\link{sedona_read_shapefile_to_typed_rdd}()},
\code{\link{sedona_save_spatial_rdd}()}
}
\concept{Sedona RDD data interface functions}
